Add rules for unpacking files
The user can specify rules for where to unpack files. Each rule has a
pattern and a destination. If the filename matches the pattern, it is
copied to the destination. Rules are only tested if the file does not
already exist in the source tree.

The init command can generate rules to put all files into the src/
directory or into src/$ext/, where $ext is the file extension. It will
also allow the user to add custom rules.