Add config command
`nasher config` can now be used to get, set, or unset configuration
options. These can be either global options that apply to all packages
or local options that apply only to a given package. These options are
separate from the package file and are not meant to be committed with
the package repository.

This change required renaming utils/config.nim to utils/options.nim.