Add source exclusion
The user can now add patterns to exclude from sources for either
packages or targets. This can make it easier to filter out unwanted
sources and can be useful to avoid repetition in target definitions. For
example, if the package already specifies a list of files and a target
wants all those files except a few, the target can inherit the package's
include list and exclude the few files it wants.