Allow spaced in user-defined compiler flags
Previously, if the user-defined nssFlags variable contained multiple
flags separated by spaces, the compiler would choke. This is a quick and
dirty fix to allow this to work.

This fix just splits the value of nssFlags on whitespace. It does this
even when the text is in quotes. This will be fixed later.