Recompile executable scripts as needed
If a script with a main() or StartingConditional() does not have a
compiled version in the script directory, it will be recompiled.
Additionally, any scripts including it will be recompiled. This mostly
fixes #6.

Known issue: if a script that includes an executable script is missing a
compiled version, it will not be recompiled unless the executable script
is recompiled also.