Implement checksums for unpack (#4)
Previously, unpacked files would overwrite the source files if they were
newer. However, this could cause problems since the unpacked files were
treated as being the age of the module file they were extracted from. If
you pack a file and then edit it in the toolset, unpacking would
overwrite any source code changes you've made since then.

This commit records checksums of the source files and module files on
unpacking. The next time the module is unpacked, each file will be
compared to the checksums to determine if the module or source files
have been updated since the last unpack and let the builder choose which
file to keep when a conflict is found.