Unpack files based on target
This changes how unpacking is done. Now, instead of specifying a file
and having it extracted to the whole source tree, you specify a target
and the file is extracted to that target's source tree. If you do not
manually specify a file, the installed version of the target's file is
unpacked. This makes the workflow as follows:

    1. Pack and install demo target: nasher install demo
    2. Edit intalled module in toolset
    3. Unpack to get toolset changes: nasher unpack demo

This should streamline the workflow and get rid of annoying prompts when
two targets have different versions of the same file: only the intended
target's file is overwritten.

Note that you can still manually specify a file to unpack:

    $ nasher unpack demo demo.mod # or...
    $ nasher unpack --file:demo.mod