Support NWN:EE module folders (closes #8)
NWN:EE added module folders. These are subdirectories of the modules
folder that contain unpacked modules. They can be loaded in the toolset
just like a packed module.

This update allows nasher to make use of this feature. It can be enabled
using the new --useModuleFolder flag. The default value for this flag is
true during installation; during unpacking, it is true if the user has
not explicity chosen a file to unpack. If you are not using NWN:EE, you
should set this flag to false using nasher config:

    $ nasher config useModuleFolder false