Improve overwrite prompt on pack and install
This adds a prompt to continue installing/launching a file when choosing
not to overwrite an existing file during pack/install. The default
answer is not to continue.

Additionally, choosing not to continue packing or installing no longer
quits the program. This means that if you are packing or installing
multiple targets with a single command, it will continue processing the
remaining targets.