Don't fail on launching non-module targets
This changes the test, play, and serve commands to continue processing
remaining targets if they encounter a non-module target. Previously, it
would error out and quit the program.

This allows you use the launch commands to install multiple targets and
then launch the module target with one command. For example, given the
targets module, hak, and tlk, running the following command would
install the hak and tlk, then install and play the module:

    nasher launch hak tlk module