Improve error handling
This change causes fatal() to throw an exception rather than just
quitting. The main module catches this and quits. It also catches other
exceptions and advises the user to report them. The goal will be for all
exceptions to eventually be gracefully handled with appropriate error
messages.

In addition, the init command now waits until unpacking to complain about
missing binaries.