Add pre-pack file filtering (#22)
This adds support for removing files from the cache before packing. This
allows you to, for example, remove `.nss` files from a hak while leaving
the compiled `.ncs` files in place. Closes #22.

To filter files, use the `filter` key under either the package or the
target. This key, like the `include` and `exclude` keys, takes a pattern
to match and can be repeated multiple times to add as many filters as
necessary. If a target does not have its own filter, it will inherit the
package-level filter. Note that filter patterns will only yield files
directly from the cache and should not contain directory information.

    [Sources]
    include = "src/*.{nss,json}"
    filter = "test_*.{nss,ncs}"

    [Target]
    name = "hak"
    file = "myhak.hak"
    filter = "*.nss"            # Will filter out all nss files
    filter = "test_*.{nss,ncs}" # Must include since not inherited

    [Target]
    name = "module"
    file = "mymod.mod"
    # No filter, so inherits the package-level filter