Compile scripts in chunks
This should fix #31. It breaks compilation into chunks to limit the
length of the command passed to the process. The new "--nssChunks" flag
controls the number of scripts that will be compiled in one pass. It
defaults to 500, but if you are still running into errors, you might try
setting it to a lower number.