Add per-target unpack rules
Any unrecognized key/value pair in the [Package] or [Target] sections
will be treated as an unpack rule of the form "pattern = destination".
If a target does not have any unpack rules of its own, it will use the
package's rules, just like it can inherit the package's include,
exclude, and filter fields.

Note that the unpack rules are only used if the file cannot be found
among the target's sources.