Use manifest to track files in cache
This fixes #20 by adding a new manifest to each target's cache. It
tracks the sha1 of the source file used to generate the cached file and
the sha1 of the cached file generated. If either of these change, the
resource will be re-converted. Using a cache separate from that used for
packing and unpacking is necessary because the user may convert or
compile without packing or unpacking.