Don't require compiler binary for some commands
Previously, all of the convert -> launch cycle commands required the
compiler binary to be present. However, some repos may not have scripts,
so the compiler check is unnecessary. This makes it so the compiler
check is only present for the compile, pack, install, play, test, and
serve commands when --noCompile is not passed. This closes #38.