Allow compiling a single file (closes #35)
This allows the `compile` command to specify a single script to compile
using the `--file` flag. This flag can take a semicolon-delimited list
or it can be repeated.

When used to compile a single file, the command will still operate on a
target. If no target is specified by the user, the default target is
used. Thus, if the script specified by the user is not in the source
tree for the target, the command will fail.

Any path information in the file is discarded, and the version of the
file in the cache is used. The reason for allowing path information in
the value is that it works with the shell's tab-completion. However,
this means that if the target's source tree contains a different version
of the file than the one the user points to, the user will get the wrong
file version. A future update will resolve this.