Fail if compiling outside source tree
Changes how compilation of single files works. Now, if only a filename
is given, it uses the file from the cache. If a source file for the
target is given, it uses the matching file from the cache. If the file
was not found in the source tree, it aborts with an error. This prevents
users from compiling the wrong version of a file.

Note that if passed with the `--noConvert` flag, the cached file could
be out of sync with the source file.