Add module name/version functionality (#51)
Thanks to @tinygiant98 for this update. This update adds the ability to set a module's name and the minimum game version required to run it. Additionally, it checks for missing .git and .gic files for areas and emits a warning to the user.

You can set the module's name and minimum game version using the modName and modMinGameVersion settings respectively. They can be passed as fields under a [Package] or [Target] section of nasher.cfg or as command-line flags. If a target does not have a set module name or version, it will inherit from the package. If neither is present, no changes will be made. A command-line option will override any setting in the nasher.cfg.