Ignore includes and exec functions in comments
This should fix #48, which caused scripts to erroneously be marked as
executable if a main() or StartingConditional() function was inside a
multi-line comment block. This had caused scripts to be recompiled on
every build because no .ncs file was generated after the last compile.

This also fixes a related error, where include directives inside
multi-line comments were erroneously detected.