Git branch feature integration (#54)
This allows targets to specify a git branch from which their files should be pulled. The `branch` parameter can be specified at the package or target level in the nasher.cfg. If a target has no branch specified, it will inherit from the package. If neither has been specified, nasher will use the currently checked-out branch. You can also override the branch on a per-command basis using the `--branch:branchname` flag.

Thanks to  @tinygiant98 for doing all the work on this. This closes #27.