Fix dockerfile. Again.
Docker builds were broken because the nim image does not add the Nimble
bin directory to PATH. This fix solves it by copying the nasher binary
to /usr/local/bin. As a side-effect, this fixes #42, allowing you to
pass your user/group to the docker command and own the files docker
generates.