feat: set NWN_ROOT and NWN_HOME before operations
When nwnsc is called with the -l flag, it first checks the NWN_ROOT
environment variable to get the location of the NWN install. This value
is also checked by the neverwinter.nim tools. If the variable has not
been set by the user, nasher will check for Steam, Beamdog, and GOG
installs and set the NWN_ROOT variable accordingly. This means users
should be able to just use the default nwnsc flags (-lowqey) and it will
find the NWN install without issue.

Similarly, the NWN_HOME directory is used for the path in which modules,
haks, etc. will be stored. If the --installDir flag was not passed, the
environment variable will be checked. If the environment variable was
not set, it will fall back to "~/Documents/Neverwinter Nights" or
"~/.local/share/Neverwinter Nights", depending on the OS.