feat: support Beamdog and GOG installs (#65)
This should allow nasher to work without additional configuration for
those who installed NWN using the Beamdog client, GOG, or non-default
Steam locations. The preferred way of setting the install location is to
set the $NWN_ROOT environment variable as this variable is also read by
nwnsc and niv's tools. In the case of nwnsc, it allows the user to run
with the default -l flag instead of having to set --nssFlags:"-n
'/path/to/nwn/data'".

When nwnsc is called with the -l flag, it first checks the NWN_ROOT
environment variable to get the location of the NWN install. This value
is also checked by the neverwinter.nim tools. If the variable has not
been set by the user, nasher will check for Steam, Beamdog, and GOG
installs and set the NWN_ROOT variable accordingly. This means users
should be able to just use the default nwnsc flags (-lowqey) and it will
find the NWN install without issue.

Similarly, the NWN_HOME directory is used for the path in which modules,
haks, etc. will be stored. If the --installDir flag was not passed, the
environment variable will be checked. If the environment variable was
not set, it will fall back to "~/Documents/Neverwinter Nights" or
"~/.local/share/Neverwinter Nights", depending on the OS.

Still needs testing on non-Linux OSes and with Beamdog installs.

When merged, will close #40.