feat: add TOML package file parsing
This is an initial proof of concept for using TOML format rather than
cfg format for package files. TOML format makes it easier to implement
subsections (e.g., [target.rules]) for better organization. It also
makes it easy to inherit fields from the package and it can account for
intentionally blanked fields.

It does away with the PackageRef type, making the Target the unit of
data. The parsing function simply returns a sequence of Targets. The
plan will be to convert the current cfg parser functions to this model.
The cfg parser will remain in place for backwards compliance, but new
additions to the system will only be supported for TOML. A utility to
convert existing cfg files to toml files may be helpful.