feat: add --onMultipleSources option
Previously, when trying to pack a target that had multiple files of the
same name within its source tree, nasher would ask the user to choose.
This was easy to miss if the `--yes` flag was passed.

This commit adds the `--onMultipleSources` option with possible values
of `choose` (manually choose a file), `default` (automatically select
the first file found), or `error` (abort with an error message). The
default value is `choose` in order to preserve the existing behavior.