feat: prompt to continue after error in chunk
This causes the user to receive a prompt if there were errors in
compilation after each chunk of scripts is run. Previously, if the user
were to compile a large number of scripts, they would have to wait until
the end of compilation to be notified. These prompts respect the
`--abortOnCompileError` flag to allow an early exit.

Note: if the `-y` flags is not in `--nssFlags`, nwnsc will stop
compiling scripts in the current chunk. If the user chooses to continue
compiling, the remaining scripts from that chunk will not have been
compiled and will show no errors to indicate why.