feat: add nasher.cfg subsections proof of concept
This is an attempt to implement subsections for the nasher.cfg file. The
current subsections allowed are [sources] and [rules]. These can either
be context-aware or explicitly set:

    # Context aware subsections
    [package]
        [sources]
    [target]
        [sources]

    # Explicit (preferred)
    [package]
        [package.sources]
    [target]
        [target.sources]

Indentation is optional, but encouraged.

The parser should be backwards compatible with existing nasher.cfg
files, and there are tests to check some known packages.

This also drops the PackageRef type, preferring instead to return a
sequence of Targets. The version, url, and authors sections are ignored
for now, but may be used again later.