feat: add user-defined variables (#25)
This allows the user to set variables that can be referenced in any
target field except for the name. You can set variables in the
[package.variables] section (to apply to all targets) or the
[target.variables] section (to apply to a single target). The key is the
variable name, while the value is what it should resolve to. The
target variables table is merged with the package variables table to
allow missing keys to be inherited.

If a variable referenced is not found, nasher will also check the
environment variables. If a variable is not found in the table or the
environment, an error is thrown.

Variables can be referenced using the syntax $variable or ${variable}.
The latter syntax allows variables to include non-alphanumeric
characters or to be adjacent to alphanumeric characters that are not
part of the variable name (e.g., ${foo}bar).

One variable is available by default: $target resolves to the name of
the current target. This allows some neat things:

  # This package has two targets, "foo" and "bar", that yield a file
  # named "foo.hak" and "bar.hak", respectively using the source files
  # in "src/foo" and "src/bar" respectively.
  [package]
  file = "$target.hak"

    [package.sources]
    include = "src/$target/*"

  [target]
  name = "foo"

  [target]
  name = "bar"

This feature can also be used to easily reference out-of-tree projects:

  [package]

    [package.variables]
    sm-utils = "../sm-utils/src" # Can be used by any target or rule

    [package.sources]
    include = "${sm-utils}/*.{nss,json}" # include files in sm-utils
    include = "src/**/*.{nss,json}"

    [package.rules]
    "util_*" = "${sm-utils}" # Unpack util files to sm-utils
    "*" = "src/$target" # Unpack unknown file to target's source dir

  [target]
  name = utils
  file = sm_utils.erf

    [target.sources]
    include = "${sm-utils}/*.{nss,json}" # include only files in sm-utils

  [target]
  name = demo
  file = core_framework.mod

    [target.variables]
    config = "src/core/core_c_config.nss" # can only be used by this target

    [target.sources]
    exclude = "${config}"