fix!: unpack multiple targets correctly
Before 0.18.0, the usage was `nasher unpack [<target> [<file>]]`. With
0.18.0, the new usage was supposed to be `nasher unpack [<target>...]`,
but a bug caused it to always use the default target.

Technically, 0.18.0 would have introduced a breaking change. This is
unbreak the breaking change, soo...

BREAKING CHANGE: If specified, the second argument to nasher unpack is
now teated as a target name rather than a file.