feat: allow targets to inherit from other targets
Previously, if a target in nasher.cfg did not have a value set, default
values could be copied from the package level. This change adds a new
keyword, `parent`, that can be added at the target level to allow a
target to copy these values from another target instead; the value
assigned to the keyword is the target to copy the values from. For
example:

  [target]
  name = "module"
  file = "my_mod.mod"

    [target.sources]
    include = "src/**/*.{nss,json}"

  [target]
  name = "module-slim"
  parent = "module"

    [target.sources]
    filter = "*.nss"

Inheirtance works the same as with inheriting package-level keys:
missing keys will be copied from the parent; existing keys will not.

Some restrictions:
1. The keyword can only be specified at the target level.
2. The parent target must be declared before the inherited target.