feat: allow user to specify default target
This adds a new nasher.cfg keyword, `default`, that can be specified at
the package or target level.

At the package level, the value is a string matching the name of the
target that should be the default target. When this target is found
during parsing, it will be added to the beginning of the target list,
making it the default target.

At the target level, this keyword does not need a value. If it is
encountered, the target will be added to the beginning of the target
list. It is possible for multiple targets to specify this keyword; each
will be added to the beginning of the list, shifting previous values
downward.

This keyword is useful because you can change the default target without
having to move things around in the nasher.cfg. It also allows the
default target to be a child of a non-default target, since parents must
be declared before children.

This example shows usage at the package level. The target `demo`
inherits the properties of `demo-nwnx`, excluding those source files
that are not desired. Because it is marked as `default`, it will be the
one acted on by commands such as `nasher pack` if no target is
specified.

  [package]
  default = "demo"

  [target]
  name = "demo-nwnx"
  description = "A demo module using nwnx"
  file = "demo.mod"

    [target.variables]
    nwnx = "lib/nwnxee"

    [target.sources]
    include = "src/**/*.{nss,json}"
    include = "${nwnx}/Core/NWScript/nwnx.nss"
    include = "${nwnx}/Plugins/Events/NWScript/nwnx_events.nss"

  [target]
  name = "demo"
  description = "A demo module without nwnx"

    [target.sources]
    exclude = "${nwnx}/**"

It is also possible to use the keyword at the target level, but this is
discouraged in all but the simplest cases since it is less obvious:

  [target]
  name = "demo-nwnx"
  description = "A demo module using nwnx"
  file = "demo.mod"

    [target.variables]
    nwnx = "lib/nwnxee"

    [target.sources]
    include = "src/**/*.{nss,json}"
    include = "${nwnx}/Core/NWScript/nwnx.nss"
    include = "${nwnx}/Plugins/Events/NWScript/nwnx_events.nss"

  [target]
  default
  name = "demo"
  description = "A demo module without nwnx"

    [target.sources]
    exclude = "${nwnx}/**"