fix: always recompile changed scripts (#104)
Previously, if a script was to be re-compiled but compilation failed on
the previous run, the script would not be re-compiled on subsequent
runs, necessitating a cache clear. This change deletes previously
compiled ncs and ndb files before compilation to ensure that, if
compilation fails, re-running the command will attempt to re-compile the
missing scripts.

This fixes #104.