feat: add skipCompile option for nasher.cfg (#118)
This adds a new directive to the `[*.sources]` section that allows skipping compilation of a file while still allowing it to be used for an include and packaged with the final module. An example use case is the Core Framework's `util_i_library.nss`, which includes a `void main()` function but is intended to be used as an include file and not executed directly. This prevents failed compilation from stopping a workflow.