fix(launch): prevent freeze on many log messages
The 1.88 preview added lots of useful log messages. However, these were
filling the output buffer when running `nasher test` or `nasher play`,
causing the game to freeze. Redirecting the logging to nasher's output
fixes the issue. However, this does result in lots of logging messages
going to nasher, which may be undesirable.